Update README.md
Processing Temperatures

Write a program that will input temperatures for 5 consecutive days. The program will store these values into an array and then average all the temperatures. It will also determine the highest temperature and the lowest temperature. The average should be displayed to two decimal places.

Don't forget Display 7-1 in the textbook has an example of finding a maximum value. Finding a minimum is similar.

Sample Run:
Input temperature 1:
68
Input temperature 2:
75
Input temperature 3:
36
Input temperature 4:
91
Input temperature 5:
84
The average temperature is 70.80
The highest temperature is 91.00
The lowest temperature is 36.00